refactor(msw): simplify solution
