+++ .github/workflows/frontend-ci.yml
@@ -158,7 +158,7 @@ jobs:
             -Dsonar.projectKey=geh-frontend-app
             -Dsonar.organization=energinet-datahub
             -Dsonar.sources=apps/dh/app-dh,libs/dh,libs/ui-watt,libs/gf
-            -Dsonar.exclusions=**/*.spec.ts,libs/dh/shared/data-access-msw/mockServiceWorker.js
+            -Dsonar.exclusions=**/*.spec.ts,libs/dh/shared/util-msw/mockServiceWorker.js
             -Dsonar.coverage.exclusions=libs/ui-watt/**/+storybook/*,libs/ui-watt/**/*.stories.ts,
         env:
           GITHUB_TOKEN: ${{ secrets.GITHUB_TOKEN }} # Needed to get PR information, if any
