+++ apps/dh/app-dh/project.json
@@ -37,7 +37,7 @@
             "output": "assets/ui-watt"
           },
           {
-            "input": "libs/dh/shared/data-access-msw",
+            "input": "libs/dh/shared/util-msw",
             "glob": "mockServiceWorker.js",
             "output": ""
           }
@@ -74,10 +74,6 @@
             {
               "replace": "libs/dh/shared/environments/src/lib/environment.ts",
               "with": "libs/dh/shared/environments/src/lib/environment.prod.ts"
-            },
-            {
-              "replace": "libs/dh/shared/data-access-msw/src/lib/msw.module.ts",
-              "with": "libs/dh/shared/data-access-msw/src/lib/msw.module.prod.ts"
             }
           ],
           "outputHashing": "all"
