+++ libs/dh/metering-point/data-access-mocks/src/index.ts
@@ -14,16 +14,4 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-import 'jest-preset-angular/setup-jest';
-
-import {
-  setUpTestbed,
-  setUpAngularTestingLibrary,
-} from '@energinet-datahub/gf/test-util-staging';
-import { addDomMatchers } from '@energinet-datahub/gf/test-util-matchers';
-
-import '@energinet-datahub/dh/shared/test-util-msw';
-
-addDomMatchers();
-setUpTestbed();
-setUpAngularTestingLibrary();
+export * from './lib/mocks';
