+++ libs/dh/metering-point/feature-search/src/test-setup.ts
@@ -23,25 +23,9 @@ import {
   setUpTestbed,
 } from '@energinet-datahub/gf/test-util-staging';
 
-// eslint-disable-next-line @nrwl/nx/enforce-module-boundaries
-import { server } from 'libs/dh/shared/environments/src/lib/mocks/server';
+import '@energinet-datahub/dh/shared/test-util-msw';
 
 addDomMatchers();
 setUpTestbed();
 setUpAngularTestingLibrary();
 setUpNgMocks();
-
-beforeAll(() => {
-  // Enable the mocking in tests.
-  server.listen();
-});
-
-afterEach(() => {
-  // Reset any runtime handlers tests may use.
-  server.resetHandlers();
-});
-
-afterAll(() => {
-  // Clean up once the tests are done.
-  server.close();
-});
