--- libs/dh/shared/data-access-msw/README.md
@@ -1,1 +1,0 @@
-# DataHub Mock Service Worker
