--- libs/dh/shared/data-access-msw/project.json
@@ -1,26 +1,0 @@
-{
-  "projectType": "library",
-  "root": "libs/dh/shared/data-access-msw",
-  "sourceRoot": "libs/dh/shared/data-access-msw/src",
-  "prefix": "dh",
-  "targets": {
-    "test": {
-      "executor": "@nrwl/jest:jest",
-      "outputs": ["coverage/libs/dh/shared/data-access-msw"],
-      "options": {
-        "jestConfig": "libs/dh/shared/data-access-msw/jest.config.js",
-        "passWithNoTests": true
-      }
-    },
-    "lint": {
-      "executor": "@nrwl/linter:eslint",
-      "options": {
-        "lintFilePatterns": [
-          "libs/dh/shared/data-access-msw/src/**/*.ts",
-          "libs/dh/shared/data-access-msw/src/**/*.html"
-        ]
-      }
-    }
-  },
-  "tags": ["product:dh", "scope:shared", "type:data-access"]
-}
