+++ libs/dh/shared/test-util-auth/src/lib/msal-fake.service.ts
@@ -63,6 +63,14 @@ function getLoggerMock(): Logger {
       console.log('error:', message);
     },
     // eslint-disable-next-line @typescript-eslint/no-unused-vars
+    info: (message: string, _correlationId?: string) => {
+      console.log('info:', message);
+    },
+    // eslint-disable-next-line @typescript-eslint/no-unused-vars
+    infoPii: (message: string, _correlationId?: string) => {
+      console.log('infoPii:', message);
+    },
+    // eslint-disable-next-line @typescript-eslint/no-unused-vars
     errorPii: (message: string, _correlationId?: string) => {
       console.log('errorPii:', message);
     },
@@ -74,5 +82,6 @@ export const MsalServiceFake = MockProvider(MsalService, {
   getLogger: getLoggerMock,
   instance: {
     getAllAccounts: () => [accountMock],
+    getActiveAccount: () => accountMock,
   } as IPublicClientApplication,
 });
