+++ libs/dh/shared/test-util-msw/README.md
@@ -1,0 +1,1 @@
+# Shared test-util mock service worker
