+++ libs/dh/shared/test-util-msw/src/lib/server.ts
@@ -1,0 +1,6 @@
+import { setupServer } from 'msw/node';
+
+import { handlers } from '@energinet-datahub/dh/shared/util-msw';
+
+// Setup requests interception using the given handlers.
+export const server = setupServer(...handlers);
