+++ libs/dh/shared/util-msw/README.md
@@ -1,0 +1,1 @@
+# Shared util Mock Service Worker
