+++ libs/dh/shared/util-msw/src/test-setup.ts
@@ -16,14 +16,6 @@
  */
 import 'jest-preset-angular/setup-jest';
 
-import {
-  setUpTestbed,
-  setUpAngularTestingLibrary,
-} from '@energinet-datahub/gf/test-util-staging';
-import { addDomMatchers } from '@energinet-datahub/gf/test-util-matchers';
+import { setUpTestbed } from '@energinet-datahub/gf/test-util-staging';
 
-import '@energinet-datahub/dh/shared/test-util-msw';
-
-addDomMatchers();
 setUpTestbed();
-setUpAngularTestingLibrary();
