+++ workspace.json
@@ -13,6 +13,7 @@
     "dh-globalization-domain": "libs/dh/globalization/domain",
     "dh-globalization-feature-language-picker": "libs/dh/globalization/feature-language-picker",
     "dh-metering-point-data-access-api": "libs/dh/metering-point/data-access-api",
+    "dh-metering-point-data-access-mocks": "libs/dh/metering-point/data-access-mocks",
     "dh-metering-point-domain": "libs/dh/metering-point/domain",
     "dh-metering-point-feature-identity-and-master-data": "libs/dh/metering-point/feature-identity-and-master-data",
     "dh-metering-point-feature-overview": "libs/dh/metering-point/feature-overview",
@@ -24,7 +25,8 @@
     "dh-metering-point-ui-status-badge": "libs/dh/metering-point/ui-status-badge",
     "dh-shared-assets": "libs/dh/shared/assets",
     "dh-shared-data-access-api": "libs/dh/shared/data-access-api",
-    "dh-shared-data-access-msw": "libs/dh/shared/data-access-msw",
+    "dh-shared-util-msw": "libs/dh/shared/util-msw",
+    "dh-shared-test-util-msw": "libs/dh/shared/test-util-msw",
     "dh-shared-environments": "libs/dh/shared/environments",
     "dh-shared-test-util-auth": "libs/dh/shared/test-util-auth",
     "dh-shared-test-util-i18n": "libs/dh/shared/test-util-i18n",
